requirements.txt
streamlit==1.31.0
tensorflow==2.15.0
numpy==1.26.4
pillow==10.2.0